requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0                    # PyTorch Image Models - state-of-the-art architectures
einops>=0.7.0                  # Tensor operations for transformers

# Data Processing & Augmentation
albumentations>=1.3.0          # Advanced augmentations (CutMix, MixUp, etc.)
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Experiment Tracking & Visualization
wandb>=0.15.0                  # Weights & Biases for experiment tracking
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration Management
hydra-core>=1.3.0              # Configuration management
omegaconf>=2.3.0

# Model Deployment
onnx>=1.14.0                   # ONNX export for deployment
onnxruntime>=1.15.0

# Utilities
tqdm>=4.66.0                   # Progress bars
pandas>=2.0.0                  # Data analysis
numpy>=1.24.0
pyyaml>=6.0

# Model Interpretability
grad-cam>=1.4.0                # Gradient-weighted Class Activation Mapping
captum>=0.6.0                  # Model interpretability

# Optimization
accelerate>=0.23.0             # Simplified multi-GPU/TPU training
transformers>=4.33.0           # For ViT and other transformer models

# Development Tools
pytest>=7.4.0
black>=23.7.0                  # Code formatting
flake8>=6.1.0                  # Linting
jupyterlab>=4.0.0             # Notebooks

# Optional: For advanced features
# efficientnet-pytorch>=0.7.0  # EfficientNet implementation
# pytorch-lightning>=2.0.0     # Alternative training framework
